Revert mergeCreature's skinWeight to Float32: 23 of 23 species were losing 6% of their vertices to the mesh origin
CROSS-LANE FIX, landed by lane 22 because it blocks the chocobo completely and
breaks the whole bestiary. `src/characters/rig/Sculpt.ts` belongs to lane 1.

`75d8768` (00:03 tonight, plan task 38) changed one row of `mergeCreature`'s
attribute table from `Float32Array` to `Uint8Array`, under a comment that said:
"Nothing in this file or in `CreatureGeo.ts` writes skin weights today, so the
row is unreachable and the guard above skips it."

The premise is false. The guard is `if (!geos[0].attributes[name]) continue`,
and `RigBuilder.attach`, `.attachBlend` and `.attachChain` all write
`skinWeight` as a `Float32BufferAttribute` (RigBuilder.ts:85, :118, :170) --
then `Rig.build` calls straight into `mergeCreature`. The row is reached by
every single creature in the bestiary. `arr.set(float32Src, off)` into a
`Uint8Array` truncates each weight toward zero.

What makes it a landmine rather than an ordinary bug is that the truncation is
selective in a way that looks like working code. A rigidly bound part writes a
weight of exactly 1.0 and survives intact; a joint-blended part writes 0.98 and
0.02 and BOTH slots become 0. A vertex whose weights sum to zero skins to the
mesh origin. So a creature keeps every `attach`ed part -- and its feet, its
beak, its eyes all render perfectly -- while everything `attachChain`ed
collapses into a fan of black triangles converging on the ground at its own
root. Nothing throws. Nothing warns.

Measured rather than argued, with a probe added here
(`src/tools/probes/skinweightblast.mts`, which builds every species prototype
and counts vertices whose weights sum below 0.5):

  23 of 23 species affected
  16,234 of 276,524 vertices (5.9%) skinning to the mesh origin
  anak 27.9%  mt 28.9%  coeurl 24.3%  dualhorn/bloodhorn 22.3%
  garula 21.0%  voretooth 18.0%  sabertusk 17.9%  arachne 15.4%
  ...and every remaining species between 0.5% and 7%.

The damage lands hardest on exactly the animals whose limbs are most blended,
which is the opposite of where anyone would look.

If a Uint8 skin-weight format is genuinely wanted for memory, it needs the
weights SCALED by 255 on the way in and `new BufferAttribute(arr, size, true)`
on the way out. An unnormalised Uint8 attribute reads 0..255 raw in the shader
and is just as wrong in the other direction. That is written into the comment
now, along with the measurement, so the next attempt starts from the numbers.

Three probes come with it: `chocobodiag` (weights, bind-pose CPU skinning and
the cloned skeleton), `chocobodiag2` (the three-stage ablation that separated
`attachChain` from `mergeCreature` -- attachChain 0 bad, attach 0 bad, merge 34
of 237 bad, and `sw.array.constructor.name === 'Uint8Array'` is the line that
ended the search), and `chocobostage` (six framings of the mount).

+++ src/characters/rig/Sculpt.ts
@@ -509,12 +509,40 @@ export function mergeCreature(list: THREE.BufferGeometry[], defMat: number[] = [
   const specs: [string, number, Float32ArrayConstructor | Uint16ArrayConstructor | Uint8ArrayConstructor][] = [
     ['position', 3, Float32Array], ['normal', 3, Float32Array], ['uv', 2, Float32Array],
     ['color', 3, Float32Array], ['aEmissive', 3, Float32Array], ['aMat', 2, Float32Array],
-    // Uint8-normalised to match `Geo.ts`'s generator (plan task 38). Nothing
-    // in this file or in `CreatureGeo.ts` writes skin weights today, so the
-    // row is unreachable and the guard above skips it — it is kept in step so
-    // that the day something does, the two builders cannot disagree about the
-    // format and hand `mergeGeometries` a null.
-    ['skinIndex', 4, Uint16Array], ['skinWeight', 4, Uint8Array],
+    /**
+     * **`skinWeight` is `Float32Array`, and it must stay that way while
+     * `RigBuilder` feeds this function.**
+     *
+     * This row was `Uint8Array` for four hours on 2026-08-31, under a comment
+     * saying "nothing in this file or in `CreatureGeo.ts` writes skin weights
+     * today, so the row is unreachable and the guard above skips it". The
+     * guard is `if (!geos[0].attributes[name]) continue`, and the premise is
+     * false: `RigBuilder.attach`, `.attachBlend` and `.attachChain` all write
+     * `skinWeight` as a `Float32BufferAttribute`, and `Rig.build` calls
+     * straight into here. The row is therefore reached by **every creature in
+     * the bestiary**, and `arr.set(float32Src, off)` into a `Uint8Array`
+     * truncates each weight toward zero.
+     *
+     * The truncation is not uniformly destructive, and that is exactly what
+     * made it invisible: a rigid `attach` writes a weight of exactly 1.0 and
+     * survives, while a blended `attachChain` writes 0.98/0.02 and both slots
+     * become 0. A vertex whose weights sum to zero skins to the mesh origin,
+     * so a creature keeps every rigidly-bound part and everything
+     * joint-blended collapses into a fan of triangles converging on its own
+     * feet. Nothing errors, and the parts that still render look correct.
+     *
+     * Measured before the revert (`src/tools/probes/skinweightblast.mts`):
+     * **23 of 23 species affected, 16,234 of 276,524 vertices** skinning to
+     * the origin -- anak 27.9%, coeurl 24.3%, dualhorn 22.3%, sabertusk 17.9%,
+     * and the collapse is worst on exactly the animals whose limbs are most
+     * blended.
+     *
+     * If a Uint8 skin-weight format is wanted for memory it needs the weights
+     * *scaled* by 255 on the way in and `new BufferAttribute(arr, size, true)`
+     * on the way out; an unnormalised Uint8 attribute reads 0..255 raw in the
+     * shader and is just as wrong in the other direction.
+     */
+    ['skinIndex', 4, Uint16Array], ['skinWeight', 4, Float32Array],
   ];
   for (const [name, size, Type] of specs) {
     if (!geos[0].attributes[name]) continue;
